Fix bug when add and delete picture in category

+++ src/Controller/Backoffice/CategoryController.php
@@ -24,13 +24,19 @@ class CategoryController extends AbstractController
     }
 
     #[Route('/new', name: 'app_category_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager): Response
+    public function new(Request $request, EntityManagerInterface $entityManager, PictureService $pictureService): Response
     {
         $category = new Category();
         $form = $this->createForm(CategoryType::class, $category);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $pictureFile = $form->get('pictures')->getData();
+            if ($pictureFile) {
+                $relativePath = $pictureService->add($pictureFile, 'categories');
+                $category->setPicture($relativePath);
+            }
+
             $entityManager->persist($category);
             $entityManager->flush();
 
@@ -39,9 +45,10 @@ class CategoryController extends AbstractController
 
         return $this->render('backoffice/category/new.html.twig', [
             'category' => $category,
-            'form' => $form,
+            'form' => $form->createView(),
         ]);
     }
+
 
     #[Route('/{id}', name: 'app_category_show', methods: ['GET'])]
     public function show(Category $category): Response
@@ -85,13 +92,25 @@ class CategoryController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_category_delete', methods: ['POST'])]
-    public function delete(Request $request, Category $category, EntityManagerInterface $entityManager): Response
+    public function delete(Request $request, Category $category, EntityManagerInterface $entityManager, PictureService $pictureService): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $category->getId(), $request->getPayload()->getString('_token'))) {
-            $entityManager->remove($category);
-            $entityManager->flush();
-        }
+        if ($this->isCsrfTokenValid('delete' . $category->getId(), $request->request->get('_token'))) {
+            // Supprimer l'image associée si elle existe
+            if ($category->getPicture()) {
+                $relativePath = $category->getPicture();
+                if ($pictureService->delete($relativePath)) {
+                } else {
+                    $this->addFlash('danger', 'La suppression de l\'image a échoué');
+                }
+            }
 
-        return $this->redirectToRoute('app_category_index', [], Response::HTTP_SEE_OTHER);
+            // Supprimer la catégorie
+            if ($this->isCsrfTokenValid('delete' . $category->getId(), $request->getPayload()->getString('_token'))) {
+                $entityManager->remove($category);
+                $entityManager->flush();
+            }
+
+            return $this->redirectToRoute('app_category_index', [], Response::HTTP_SEE_OTHER);
+        }
     }
 }
